#I4-publish command Удалил serdes файлы и перенс весь код в модуль serdes. Исправил замечания clippy. Отключил must_use_candidate.
