+++ src/command/index.rs
@@ -1,10 +1,10 @@
-use crate::commands::{CommandResult, Error, MashinkaCommand, INDEX_COMMAND_NAME};
+use crate::command::{Command, CommandResult, Error, INDEX_COMMAND_NAME};
 use std::collections::HashMap;
 
-pub struct IndexCommand;
+pub struct Index;
 
 /// Индексирует записи для поиска
-impl MashinkaCommand for IndexCommand {
+impl Command for Index {
     fn run(&self) -> Result<CommandResult, Error> {
         Ok(CommandResult {
             command: INDEX_COMMAND_NAME.to_string(),
